test: :sparkles: update tests and rename files for clarity
With the update to `ExtractRouteData`, the tests were adjusted to reflect the new behavior where query parameters are optional. Additionally, file names were renamed for better clarity: - `path_with_param.test.ts` -> `path_with_paramts.test.ts` - `path_with_query.test.ts` -> `path_with_queries.test.ts`  These changes ensure that tests are aligned with the updated logic and that file names clearly indicate their purpose.
